Formalize reviews hook by moving into hooks.ts
